casper-storage/trie store cache: avoid serializing trie twice on store
Trie nodes are written to storage by serializing them with bytesrepr and
inserting them under the key generated by hashing the serialized
representation.
The `trie_hash` method of the `Trie` was serializing the struct to
calculate the hash and then discarded the bytes. But we also need to
serialize the Trie in order to store it to disk.
We now re-use the bytes instead of discarding them. This adds a further
10% improvement in batch insert speed to the cache.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>