casper-storage: use batch write to commit data to store
The upgrade code uses a tracking copy to perform the global state
upgrade. This populates a cache with the values that are ultimatelly
going to be written to the store. The tracking copy also records the
transform log in the form of `Effects`.
In order to actually write the values to the store, the tracking copy
can return the effects and the upgrade code used to apply those effects
to the store through the `commit` API.
This is inneficient because we need to go through all of the effects one
by one and apply them to the store.
In order to improve this process, a new api called `commit_values` is
provided that allows the upgrade code to store (key, value) pairs
directly.
We get the pairs that we want to store and the keys that we want to
prune by consuming and destructuring the tracking copy. As a bonus, we
also prevent cloning the effects without need.

The tracking copy looks a lot like the Scratch store at this point. In
the future we might want to remove the scratch store altogether since it
already doesn't have a very consistent API (e.g. committing does not
produce a new state hash).
In the future we might want to do a biger refactor in order to allow a
TrackingCopy to also use a `StoreWriter` interface that can also flush
the data to the store.
However introducing this `commit_values` API now allows us to quickly
realize the benefits of the batch write optimizations without a large
re-write.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>